# docs/swagger.yaml
definitions:
  auth.AddPermsReq:
    properties:
      permission_ids:
        items:
          type: string
        type: array
      role_id:
        type: integer
    required:
    - permission_ids
    - role_id
    type: object
  auth.CreateRoleReq:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  auth.DelPermsReq:
    properties:
      permission_ids:
        items:
          type: string
        type: array
      role_id:
        type: integer
    required:
    - permission_ids
    - role_id
    type: object
  auth.DeleteRoleReq:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  auth.UpdateRoleReq:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - id
    - name
    type: object
  login.LoginReq:
    properties:
      password:
        type: string
      phone:
        type: string
    required:
    - password
    - phone
    type: object
  menu.CreateMenuReq:
    properties:
      icon:
        type: string
      label:
        type: string
      parent_id:
        type: integer
      permissions:
        type: string
      target:
        type: string
      to:
        type: string
    required:
    - label
    type: object
  menu.DeleteMenuReq:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  menu.UpdateMenuReq:
    properties:
      icon:
        type: string
      id:
        type: integer
      label:
        type: string
      parent_id:
        type: integer
      permissions:
        type: string
      target:
        type: string
      to:
        type: string
    required:
    - id
    - label
    type: object
  model.Menu:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      icon:
        type: string
      id:
        type: integer
      label:
        type: string
      parent_id:
        type: integer
      permissions:
        type: string
      target:
        type: string
      to:
        type: string
      updated_at:
        type: string
    type: object
  model.Permission:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      desc:
        type: string
      id:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  model.RelUserRole:
    properties:
      id:
        type: integer
      role_id:
        type: integer
      user_id:
        type: integer
    type: object
  model.Role:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  page.CreatePageReq:
    properties:
      config:
        type: string
      extend:
        type: string
      icon:
        type: string
      name:
        type: string
      path:
        type: string
    required:
    - config
    - name
    - path
    type: object
  page.DeletePageReq:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  page.UpdatePageReq:
    properties:
      config:
        type: string
      extend:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      path:
        type: string
    required:
    - config
    - id
    - name
    - path
    type: object
  user.AddUserRolesReq:
    properties:
      role_id:
        items:
          type: integer
        type: array
      user_id:
        type: integer
    required:
    - role_id
    - user_id
    type: object
  user.DelUserRolesReq:
    properties:
      role_id:
        items:
          type: integer
        type: array
      user_id:
        type: integer
    required:
    - role_id
    - user_id
    type: object
info:
  contact: {}
  title: API 文档
  version: "1.0"
paths:
  /api/config/page:
    get:
      parameters:
      - description: path
        in: query
        name: path
        required: true
        type: string
      responses: {}
      summary: 页面详情
      tags:
      - 页面
  /api/login/adminLoginInfo:
    get:
      responses: {}
      summary: 获取管理员登录信息
      tags:
      - 登录
  /api/login/loginByPhone:
    post:
      parameters:
      - description: body
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/login.LoginReq'
      responses: {}
      summary: 手机号密码登录
      tags:
      - 登录
  /api/login/logout:
    get:
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 注销
      tags:
      - 登录
  /api/menu/create:
    post:
      parameters:
      - description: 参数
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/menu.CreateMenuReq'
      responses:
        "200":
          description: OK
          schema:
            type: number
      summary: 创建菜单
      tags:
      - 菜单
  /api/menu/delete:
    post:
      parameters:
      - description: 参数
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/menu.DeleteMenuReq'
      responses:
        "200":
          description: OK
          schema:
            type: number
      summary: 删除菜单
      tags:
      - 菜单
  /api/menu/list:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Menu'
      summary: 菜单列表
      tags:
      - 菜单
  /api/menu/update:
    post:
      parameters:
      - description: 参数
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/menu.UpdateMenuReq'
      responses:
        "200":
          description: OK
          schema:
            type: number
      summary: 更新菜单
      tags:
      - 菜单
  /api/page/create:
    post:
      parameters:
      - description: 参数
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/page.CreatePageReq'
      responses:
        "200":
          description: OK
          schema:
            type: number
      summary: 创建页面
      tags:
      - 页面
  /api/page/delete:
    post:
      parameters:
      - description: 参数
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/page.DeletePageReq'
      responses:
        "200":
          description: OK
          schema:
            type: number
      summary: 删除页面
      tags:
      - 页面
  /api/page/list:
    get:
      parameters:
      - in: query
        name: name
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: path
        type: string
      - in: query
        name: per_page
        type: integer
      responses: {}
      summary: 页面列表
      tags:
      - 页面
  /api/page/update:
    post:
      parameters:
      - description: 参数
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/page.UpdatePageReq'
      responses:
        "200":
          description: OK
          schema:
            type: number
      summary: 更新页面
      tags:
      - 页面
  /api/perm/list:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Permission'
      summary: 权限列表
      tags:
      - 权限
  /api/role/addPerms:
    post:
      parameters:
      - description: 参数
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/auth.AddPermsReq'
      responses:
        "200":
          description: OK
          schema:
            type: number
      summary: 为角色添加权限
      tags:
      - 角色
  /api/role/create:
    post:
      parameters:
      - description: 参数
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/auth.CreateRoleReq'
      responses:
        "200":
          description: OK
          schema:
            type: number
      summary: 创建角色
      tags:
      - 角色
  /api/role/delPerms:
    post:
      parameters:
      - description: 参数
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/auth.DelPermsReq'
      responses:
        "200":
          description: OK
          schema:
            type: number
      summary: 为角色移除权限
      tags:
      - 角色
  /api/role/delete:
    post:
      parameters:
      - description: 参数
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/auth.DeleteRoleReq'
      responses:
        "200":
          description: OK
          schema:
            type: number
      summary: 删除角色
      tags:
      - 角色
  /api/role/detail/{id}:
    get:
      parameters:
      - description: 角色 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Role'
      summary: 根据id获取角色
      tags:
      - 角色
  /api/role/getRolePerms:
    get:
      parameters:
      - in: query
        name: role_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Permission'
      summary: 根据角色id获取角色权限
      tags:
      - 角色
  /api/role/list:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Role'
      summary: 角色列表
      tags:
      - 角色
  /api/role/update:
    post:
      parameters:
      - description: 参数
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/auth.UpdateRoleReq'
      responses:
        "200":
          description: OK
          schema:
            type: number
      summary: 更新角色
      tags:
      - 角色
  /api/user/addUserRoles:
    post:
      parameters:
      - description: 参数
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/user.AddUserRolesReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.RelUserRole'
      summary: 添加用户角色
      tags:
      - 用户
  /api/user/delUserRoles:
    post:
      parameters:
      - description: 参数
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/user.DelUserRolesReq'
      responses:
        "200":
          description: OK
          schema:
            type: number
      summary: 为用户移除角色
      tags:
      - 用户
swagger: "2.0"